[dev-2992] removing duplicate code ... used for getting tooltip props
