+++ src/js/components/awardv2/shared/awardAmountsSection/charts/ExceedsCurrentChart.jsx
@@ -9,16 +9,8 @@ import PropTypes from 'prop-types';
 import { generatePercentage } from 'helpers/aggregatedAmountsHelper';
 
 import TooltipWrapper from "../../../../sharedComponents/TooltipWrapper";
-import {
-    CombinedObligatedAmounts,
-    CombinedCurrentAmounts,
-    CombinedPotentialAmounts,
-    CombinedExceedsCurrentAmounts,
-    ObligatedAmountTooltip,
-    CurrentAmountTooltip,
-    PotentialAmountTooltip,
-    ExceedsCurrentAmountTooltip
-} from "../Tooltips";
+import { getTooltipPropsByAwardTypeAndSpendingCategory } from '../Tooltips';
+
 import { useTooltips } from "./AwardAmountsChart";
 import { AWARD_V2_AGGREGATED_AMOUNTS_PROPS, TOOLTIP_PROPS } from '../../../../../propTypes/index';
 
@@ -32,49 +24,6 @@ const propTypes = {
     exceedsCurrentTooltipProps: TOOLTIP_PROPS
 };
 
-const tooltipPropsByAwardTypeAndSpendingCategory = (type, category, data) => {
-    const map = {
-        idv: {
-            obligated: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedObligatedAmounts total={data.obligatedFormatted} count={data.childAwardCount} />
-            },
-            current: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedCurrentAmounts total={data.baseAndExercisedOptionsFormatted} count={data.childAwardCount} />
-            },
-            potential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedPotentialAmounts total={data.baseAndAllOptionsFormatted} count={data.childAwardCount} />
-            },
-            exceedsCurrent: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedExceedsCurrentAmounts total={data.overspendingFormatted} count={data.childAwardCount} />
-            }
-        },
-        contract: {
-            obligated: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <ObligatedAmountTooltip />
-            },
-            current: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CurrentAmountTooltip />
-            },
-            potential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <PotentialAmountTooltip />
-            },
-            exceedsCurrent: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <ExceedsCurrentAmountTooltip />
-            }
-        }
-    };
-
-    return map[type][category];
-};
-
 const ExceedsCurrentChart = ({ awardAmounts, awardType }) => {
     const [
         activeTooltip,
@@ -86,7 +35,7 @@ const ExceedsCurrentChart = ({ awardAmounts, awardType }) => {
     ] = useTooltips(["obligated", "current", "potential", "exceedsCurrent"]);
 
     const buildTooltipProps = (spendingCategory, isVisible, showTooltip, type = awardType) => ({
-        ...tooltipPropsByAwardTypeAndSpendingCategory(type, spendingCategory, awardAmounts),
+        ...getTooltipPropsByAwardTypeAndSpendingCategory(type, spendingCategory, awardAmounts),
         wide: true,
         controlledProps: {
             isControlled: true,
