+++ src/js/components/awardv2/shared/awardAmountsSection/charts/ExceedsPotentialChart.jsx
@@ -8,16 +8,8 @@ import PropTypes from 'prop-types';
 
 import { generatePercentage } from 'helpers/aggregatedAmountsHelper';
 import TooltipWrapper from "../../../../sharedComponents/TooltipWrapper";
-import {
-    CombinedObligatedAmounts,
-    CombinedCurrentAmounts,
-    CombinedPotentialAmounts,
-    CombinedExceedsPotentialAmounts,
-    ObligatedAmountTooltip,
-    CurrentAmountTooltip,
-    PotentialAmountTooltip,
-    ExceedsPotentialAmountTooltip
-} from "../Tooltips";
+import { getTooltipPropsByAwardTypeAndSpendingCategory } from '../Tooltips';
+
 import { useTooltips } from "./AwardAmountsChart";
 import { AWARD_V2_AGGREGATED_AMOUNTS_PROPS, TOOLTIP_PROPS } from '../../../../../propTypes/index';
 
@@ -30,48 +22,6 @@ const propTypes = {
     exceedsPotentialTooltipProps: TOOLTIP_PROPS
 };
 
-const tooltipPropsByAwardTypeAndSpendingCategory = (type, category, data) => {
-    const map = {
-        idv: {
-            obligated: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedObligatedAmounts total={data.obligatedFormatted} count={data.childAwardCount} />
-            },
-            current: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedCurrentAmounts total={data.baseAndExercisedOptionsFormatted} count={data.childAwardCount} />
-            },
-            potential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedPotentialAmounts total={data.baseAndAllOptionsFormatted} count={data.childAwardCount} />
-            },
-            exceedsPotential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CombinedExceedsPotentialAmounts total={data.extremeOverspendingFormatted} count={data.childAwardCount} />
-            }
-        },
-        contract: {
-            obligated: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <ObligatedAmountTooltip />
-            },
-            current: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <CurrentAmountTooltip />
-            },
-            potential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <PotentialAmountTooltip />
-            },
-            exceedsPotential: {
-                offsetAdjustments: { top: 0 },
-                tooltipComponent: <ExceedsPotentialAmountTooltip />
-            }
-        }
-    };
-
-    return map[type][category];
-};
 const ExceedsPotentialChart = ({ awardType, awardAmounts }) => {
     const [
         activeTooltip,
@@ -83,7 +33,7 @@ const ExceedsPotentialChart = ({ awardType, awardAmounts }) => {
     ] = useTooltips(["obligated", "current", "potential", "exceedsPotential"]);
 
     const buildTooltipProps = (spendingCategory, isVisible, showTooltip, type = awardType) => ({
-        ...tooltipPropsByAwardTypeAndSpendingCategory(type, spendingCategory, awardAmounts),
+        ...getTooltipPropsByAwardTypeAndSpendingCategory(type, spendingCategory, awardAmounts),
         wide: true,
         controlledProps: {
             isControlled: true,
